requirements.txt
pywebview>=4.4.1
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.3.0
pyautogui>=0.9.54
keyboard>=0.13.5
pyinstaller>=6.6.0
